requirements.txt
telethon
python-dotenv
sqlalchemy
gspread
google-auth
